+++ web_ui/src/AppSettings.js
@@ -29,15 +29,18 @@ const AppSettings = ({ appSettingsOpen, setAppSettingsOpen, user, setUser, onClo
     const handleResize = useCallback( 
         // Slow down resize events to avoid excessive re-rendering and avoid ResizeObserver loop limit exceeded error
         debounce((entries) => {
-        const { width } = entries[0].contentRect;
-        setWidth(width);
+            entries && entries.length > 0 && setWidth(entries[0].contentRect.width);
         }, 100),
         []
     );
 
     useEffect(() => {
         const element = document.getElementById("chat-panel");
-        const observer = new ResizeObserver(handleResize);
+        const observer = new ResizeObserver((entries) => {
+            if (entries && entries.length > 0 && entries[0].target === element) {
+              handleResize();
+            }
+        });
         element && observer.observe(element);
         return () => observer.disconnect();
     }, [handleResize]);